Move ng to a top level directory
ng can be served independently from gsad daemon (as it is done with the
development server already). Therefore move ng to the top level
directory and the sources for gsad into a new gsad directory. This
separates ng and gsad cleanly.

With this change the strange src/html/classic/ng/src path is removed
too.

Adjust doxygen files to use gsad/ as source directory.