Remove optional use of xsltproc command line tool.
The code had an optional method to use the xsltproc command line tool.
By default it was not used and rather the xslt library was used.
This patch make the use of the library mandatory and removes any
code path for the xsltproc tool.

Note that the tool is still in use for build-time to check validity
of the installed xsl-files.

* gsad/CMakeLists.txt: Drop handling of USE_LIBXSLT. Since it was the
  last element to determine about MANDATORY_TOOL_MISSING, the handling
  of this variable is also removed entirely.

* gsad/src/CMakeLists.txt: Drop handling og USE_LIBXSLT.

* gsad/src/gsad.c (register_signal_handlers): Make the code path for
  USE_LIBXSLT the only one.

* gsad/src/gsad_base.c (gsad_base_init, gsad_base_cleanup,
  xsl_transform_with_stylesheet): Make the code path for
  USE_LIBXSLT the only one.