Fix validation of delta_report_id
The report_id got changed to be an alias of id and moved below the
definition of the delta_report_id parameter. Therefore delta_report_id
valiator couldn't reference the report_id alias and actually always
failed the validation of the paramter. As a result delta reports
couldn't be requested and loaded from gvmd.