Use params_value() for allow_simultaneous_ips
Although the underlying value is bool we should
not use params_value_bool for retrieving the
value of allow_simultaneous_ips because it would
break the mechanism of setting a default value
for the setting.

params_value() returns a string ("0" or "1") if
there is an value and NULL if not.
params_value_bool() return 0 if the value is "0"
OR there is no value.
This means that if we want to be able to set a
default if there is no value (NULL) we need to use
params_value() else the information is lost.

(cherry picked from commit 0bcf38260aad50ea2bc0990ec3ab25b5d47b9954)