Avoid caching of the index.html page
When updating to a newer GSA version the client's browser may still have
an old version of that index.html file in its cache. In that case it
wont load the new JavaScript files. Therefore don't cache all URLs where
the index.html file is returned. This ensures the browers always loads
the newest JavaScript files.