Fix: Fix regex for new glib2.0 >= 2.73.2
Patch by Sophie Brun <sophie@offensive-security.com>

Since glib2.0 version 2.73.2, they replace PCRE1 with PCRE2 and it seems
that some functionalities are no longer supported.
See https://gitlab.gnome.org/GNOME/glib/-/issues/2760
http://www.pcre.org/current/doc/html/pcre2pattern.html#SEC9
We need to place the hyphen at the end of the class.
" If a minus character is required in a class, it must be escaped with a
backslash or appear in a position where it cannot be interpreted as
indicating a range, typically as the first or last character in the
class, or immediately after a range."

Bug-Kali: https://bugs.kali.org/view.php?id=7926