Change: Use different return values for gvm_validate
Improve the return values of gvm_validate for the different error cases.
Normally it should only return 0 if a regexp matches or 2 if not. All
other return values are error cases. Would be nice to use 1 for a not
matching regexp but without having unit tests for all gsad functions
changing this may introduce regressions.