Change: Handle manual URL as static content
For static content an index.html is returned if available when a
directory is requested. If the URL points to a directory but the URL
doesn't end with a slash a redirect response with a URL including the
slash is returned. With this change gsad behaves more like a "normal"
http server when serving static content for the manual.